refactor: view 기본 파일 이름 변경 및 경로수정
- view.js -> index.js
